+++ src/main/kotlin/com/atlassian/performance/tools/concurrency/api/TaskContext.kt
@@ -1,8 +1,9 @@
 package com.atlassian.performance.tools.concurrency.api
 
-import com.atlassian.performance.tools.jvmtasks.api.TaskTimer
+import com.atlassian.performance.tools.jvmtasks.api.TaskScope
 import org.apache.logging.log4j.CloseableThreadContext
 import org.apache.logging.log4j.ThreadContext
+import java.util.concurrent.Callable
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.ExecutorService
 import java.util.function.Supplier
@@ -21,7 +22,7 @@ fun <T> ExecutorService.submitWithLogContext(
         Supplier {
             CloseableThreadContext.pushAll(logStack.asList())
                 .putAll(logContext)
-                .use { TaskTimer.time(label, task) }
+                .use { TaskScope.task(label, Callable(task)) }
         },
         this
     )