+++ CHANGELOG.md
@@ -34,6 +34,7 @@ E.g. you can depend on `[1.3.0, 2.0.0)` and know that `log4j-core` will not come
 
 ### Fixed
 - Relax `log4j-api` dependency to a SemVer range.
+- Use `TaskScope` instead of `TaskTimer`.
 
 ## [1.2.1] - 2023-03-21
 [1.2.1]: https://github.com/atlassian/concurrency/compare/release-1.2.0...release-1.2.1